+++ client/src/components/book-list/PointingFinger.tsx
